AE-379: Addressing pull request comments, fixing issue with Jest tests leaving async functions unhandled.
