+++ src/compare-oracle-to-fallback/compare-oracle-to-fallback.spec.js
@@ -1,54 +1,54 @@
 // required libraries
 const BigNumber = require('bignumber.js');
 const { createBlockEvent } = require('forta-agent');
-const { 
-  provideHandleBlock,
-  createAlert,
-  calculatePercentError
-} = require('./compare-oracle-to-fallback');
+
+// mock the initializeTokensContractsAlerts async function before importing the agent module
+// the agent module will then receive our mocked version of the function when it is imported
+jest.mock('./agent-setup', () => ({
+  ...jest.requireActual('./agent-setup'),
+  initializeTokensContractsAlerts: jest.fn().mockResolvedValue(),
+}));
+
+const { createAlert, calculatePercentError } = require('./agent-setup');
+const { provideHandleBlock } = require('./compare-oracle-to-fallback');
 
 const SECONDS_PER_DAY = 86400;
 
 describe('AAVE oracle versus fallback oracle agent', () => {
   let handleBlock;
 
-  function getMockFunction(assetPriceOracle, assetPriceFallback, tStart) {
-    // this function allows us to mock the behavior of the initializeTokensContracts function
-    // in production, that function:
-    return function mockTokensContractsAlertsPromise() {
-      // create a token
-      const reserveToken = { symbol: 'FAKE1', tokenAddress: '0xFIRSTFAKETOKENADDRESS' };
+  function mockTokensContractsAlertsPromise(assetPriceOracle, assetPriceFallback, tStart) {
+    // create a token
+    const reserveToken = { symbol: 'FAKE1', tokenAddress: '0xFIRSTFAKETOKENADDRESS' };
 
-      // create a mock price oracle contract for returning the value we want to test with
-      const priceOracleContractInstance = {
-        getAssetPrice: jest.fn(() => Promise.resolve(assetPriceOracle)),
-      };
-
-      // create a mock fallback oracle contract for returning the value we want to test with
-      const fallbackOracleContractInstance = {
-        getAssetPrice: jest.fn(() => Promise.resolve(assetPriceFallback)),
-      };
-
-      // create an object to track the number of token alerts in the last day
-      const tokenAlert = { numAlertsInLastDay: 0, tStart };
-
-      // create the Array of Tuples of tokens / oracle / fallback / alerts objects
-      const tokenContractFallbackAlertTuples = [
-        {
-          reserveToken, priceOracleContractInstance, fallbackOracleContractInstance, tokenAlert,
-        },
-      ];
-
-      // wrap the Array of Tuples in a Promise (which we will resolve immediately)
-      // the original function returns a Promise that is resolved when all contract interactions
-      // have finished
-      return Promise.resolve(tokenContractFallbackAlertTuples);
+    // create a mock price oracle contract for returning the value we want to test with
+    const aaveOracleContractInstance = {
+      getAssetPrice: jest.fn(() => Promise.resolve(assetPriceOracle)),
     };
+
+    // create a mock fallback oracle contract for returning the value we want to test with
+    const fallbackOracleContractInstance = {
+      getAssetPrice: jest.fn(() => Promise.resolve(assetPriceFallback)),
+    };
+
+    // create an object to track the number of token alerts in the last day
+    const tokenAlert = { numAlertsInLastDay: 0, tStart };
+
+    // create the Array of Tuples of tokens / oracle / fallback / alerts objects
+    const tokenContractFallbackAlertTuples = [
+      {
+        reserveToken, aaveOracleContractInstance, fallbackOracleContractInstance, tokenAlert,
+      },
+    ];
+
+    // wrap the Array of Tuples in a Promise (which we will resolve immediately)
+    // the original function returns a Promise that is resolved when all contract interactions
+    // have finished
+    return Promise.resolve(tokenContractFallbackAlertTuples);
   }
 
   describe('Fallback Price Oracle Monitoring', () => {
-    it('returns findings if percent error between oracle and fallback is > 2%'
-      + ' and last alert was created more than 24 hours ago', async () => {
+    it('returns findings if percent error between oracle and fallback is > 2% and last alert was created more than 24 hours ago', async () => {
       // create an oracle asset price an a fallback oracle price that will trigger an alert
       const assetPriceOracle = new BigNumber(100);
       const assetPriceFallback = new BigNumber(97.9);
@@ -68,13 +68,17 @@ describe('AAVE oracle versus fallback oracle agent', () => {
       });
 
       // create the mocked promise to pass in to provideHandleBlock
-      const tokensContractsAlerts = getMockFunction(assetPriceOracle, assetPriceFallback, tStart);
+      const tokensContractsAlerts = await mockTokensContractsAlertsPromise(
+        assetPriceOracle,
+        assetPriceFallback,
+        tStart
+      );
 
       // create the block handler
       handleBlock = provideHandleBlock(tokensContractsAlerts);
 
       // create expected finding
-      const { reserveToken } = (await tokensContractsAlerts())[0];
+      const { reserveToken } = tokensContractsAlerts[0];
       const percentError = calculatePercentError(assetPriceOracle, assetPriceFallback);
       const numAlertsInLastDay = 0;
 
@@ -110,7 +114,11 @@ describe('AAVE oracle versus fallback oracle agent', () => {
       });
 
       // create the mocked promise to pass in to provideHandleBlock
-      const tokensContractsAlerts = getMockFunction(assetPriceOracle, assetPriceFallback, tStart);
+      const tokensContractsAlerts = await mockTokensContractsAlertsPromise(
+        assetPriceOracle,
+        assetPriceFallback,
+        tStart
+      );
 
       // create the block handler
       handleBlock = provideHandleBlock(tokensContractsAlerts);
@@ -120,7 +128,7 @@ describe('AAVE oracle versus fallback oracle agent', () => {
 
       // we also expect that the variable tracking the number of alerts in the last 24 hours will
       // be incremented by 1
-      const { tokenAlert } = (await tokensContractsAlerts())[0];
+      const { tokenAlert } = tokensContractsAlerts[0];
       expect(tokenAlert.numAlertsInLastDay).toStrictEqual(1);
     });
 
@@ -144,7 +152,11 @@ describe('AAVE oracle versus fallback oracle agent', () => {
       });
 
       // create the mocked promise to pass in to provideHandleBlock
-      const tokensContractsAlerts = getMockFunction(assetPriceOracle, assetPriceFallback, tStart);
+      const tokensContractsAlerts = await mockTokensContractsAlertsPromise(
+        assetPriceOracle,
+        assetPriceFallback,
+        tStart
+      );
 
       // create the block handler
       handleBlock = provideHandleBlock(tokensContractsAlerts);
